add common styles
