--- website/frontend/src/components/pages/Messages/Messages.scss
@@ -1,33 +1,0 @@
-@import 'assets/scss/variables';
-
-.messages {
-  .pagination__wrapper {
-    margin: 26px 0;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    .pagination__wrapper-caption {
-      font-weight: 300;
-      font-size: 21px;
-      line-height: 130%;
-      display: flex;
-      align-items: center;
-      color: $textColorSecondary;
-    }
-    button {
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      gap: 10px;
-      font-weight: bold;
-      font-size: 16px;
-      line-height: 24px;
-      color: #ffffff;
-    }
-  }
-  .pagination_bottom {
-    margin-bottom: 76px;
-    margin-top: 24px;
-  }
-}
