+++ website/frontend/src/components/pages/Programs/Programs.module.scss
@@ -3,3 +3,23 @@
 .main {
   margin: $desktopSidesMargin;
 }
+
+.topPagination {
+  display: flex;
+  margin: 20px 0;
+  justify-content: space-between;
+  align-items: center;
+
+  .caption {
+    display: flex;
+    font-size: $fontSizeLarge;
+    line-height: 1.3;
+    font-weight: 300;
+    align-items: center;
+    color: $textColorSecondary;
+  }
+}
+
+.bottomPagination {
+  margin-top: 20px;
+}
