--- website/frontend/src/components/pages/Programs/children/All/All.module.scss
@@ -1,42 +1,0 @@
-@use 'assets/scss/variables' as *;
-@use 'assets/scss/breakpoints.scss' as *;
-
-.paginationWrapper {
-  margin: 26px 0;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-
-  div:first-child {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 44px;
-  }
-
-  button {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 10px;
-    font-weight: bold;
-    font-size: 16px;
-    line-height: 24px;
-    color: $textColor;
-  }
-
-  span {
-    font-weight: 300;
-    font-size: 21px;
-    line-height: 130%;
-    display: flex;
-    align-items: center;
-    color: $textColorSecondary;
-  }
-}
-
-.allProgramsList {
-  display: flex;
-  flex-direction: column;
-}
