+++ website/frontend/src/components/pages/Programs/children/All/index.ts
@@ -1,0 +1,1 @@
+export * from './All';
