+++ website/frontend/src/components/pages/Programs/children/Messages/index.ts
@@ -1,0 +1,1 @@
+export * from './Messages';
