--- website/frontend/src/components/pages/Programs/children/Recent/Recent.module.scss
@@ -1,37 +1,0 @@
-@use 'assets/scss/variables' as *;
-
-.noMessage {
-  color: $textColorSecondary;
-  font-size: 18px;
-  font-weight: 600;
-  margin-bottom: 40px;
-  text-transform: uppercase;
-  text-align: center;
-}
-.paginationWrapper {
-  margin: 26px 0;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-
-  div:first-child {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 44px;
-  }
-
-  span {
-    font-weight: 300;
-    font-size: 21px;
-    line-height: 130%;
-    display: flex;
-    align-items: center;
-    color: #858585;
-  }
-}
-.paginationBottom {
-  margin-bottom: 76px;
-  margin-top: 24px;
-}
