+++ website/frontend/src/components/pages/Programs/children/Recent/index.ts
@@ -1,0 +1,1 @@
+export * from './Recent';
