+++ website/frontend/src/components/pages/Programs/children/Upload/index.ts
@@ -1,0 +1,1 @@
+export * from './Upload';
